Add visual mesh URI resolver test
